pyproject.toml
[tool.poetry]
name = "paper-assistant"
version = "0.1.0"
description = ""
authors = ["Ahmed Hatem <ahmed-1998abdallah@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
pandas = "^2.1.1"
numpy = "^1.26.1"
ipykernel = "^6.25.2"
langchain = "^0.0.320"
llama-cpp-python = "0.2.11"
sentence-transformers = "^2.2.2"
fastapi = "^0.104.0"
pydantic = "^2.4.2"
pre-commit = "^3.5.0"
black = "^23.11.0"
flake8 = "^6.1.0"
isort = "^5.12.0"
openai = "^1.3.5"
python-dotenv = "^1.0.0"
uvicorn = "^0.25.0"
chromadb = "^0.4.21"
lxml = "^5.0.0"
tiktoken = "^0.5.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
target-version = ['py39']
experimental-string-processing = true
[tool.isort]
profile = "black"
multi_line_output = 3
py_version = 27